Add options to turn off haskellmode's completion.
This patch adds options for turning off haskellmode's completion since
it currently tramples over the omnifunc (with GHC_CompleteImports) and
completefunc (with CompleteHaddock), regardless of whether it is already
set. This is a problem if the user wants to set omnifunc/completefunc to
something else (e.g. if the user is using necoghc or an autocompletion
system like YCM or neocomplete). The names of the added options are
g:haskellmode_completion_ghc and g:haskellmode_completion_haddock.